feat(menu): link to en-US if page not available in current locale (#547)
* feat(menu): link to en-US if translation missing

* enhance(menu): render doc links via helper

* enhance(menu): show en-US indicator for cross-locale links

* chore(menu): add script to check missing docs

* fix(menu): add en-US indicator for curriculum

For the other links with icon, it is less trivial, because
those are currently using an inline-flex with a gap, which
would cause the indicator to be vertically centered, and
have an unexpected gap next to them.

We also cannot add the indicator behind Blog, because there is
not necessarily enough space, so this might break the layout.

* Apply suggestions from code review

Co-authored-by: Leo McArdle <leo@mca.is>

---------

Co-authored-by: Leo McArdle <leo@mca.is>